Attempt to be compatible with newer versions of Docker
Previously, nvidia-docker aborted if it detected a version of Docker
more recent than the one it supports. In hindsight, this provided
little benefits since the list of flags has now become quite stable,
and it forced our users to frequently update nvidia-docker manually,
because nvidia-docker is not available on a public package repository.

This problem is exacerbated now that Docker is on a monthly release schedule:
https://github.com/docker/docker/pull/31075
We don't want to generate new packages every month.

Thus, if the version of Docker detected by nvidia-docker is newer than
its latest known Docker version, we will use the boolean flags of this
latest known version.