Convert nvidia-docker2 to a meta package
This change converts nvidia-docker2 to a meta package that serves
only to introduce a dependency on the nvidia-container-toolkit
package. This allows workflows that expect nvidia-docker2 to be
installed to continue to function.

Note that this also removes the dependency on various docker packages.

The nvidia-contianer-toolkit-base package includes the

nvidia-ctk runtime configure

command that can be used to apply the required config to the docker
daemon.json file instead of overwriting it entirely.

Signed-off-by: Evan Lezar <elezar@nvidia.com>